Move DomainConfigs around, add doc comments
