+++ src/api/main/MainLocator.ts
@@ -4,7 +4,7 @@ import { EventController } from "./EventController"
 import { EntropyCollector } from "./EntropyCollector"
 import { SearchModel } from "../../search/model/SearchModel"
 import { MailboxDetail, MailModel } from "../../mail/model/MailModel"
-import { assertMainOrNode, DomainConfigProvider, isAndroidApp, isApp, isBrowser, isDesktop, isElectronClient, isIOSApp } from "../common/Env"
+import { assertMainOrNode, isAndroidApp, isApp, isBrowser, isDesktop, isElectronClient, isIOSApp } from "../common/Env"
 import { notifications } from "../../gui/Notifications"
 import { LoginController } from "./LoginController"
 import type { ContactModel } from "../../contacts/model/ContactModel"
@@ -98,6 +98,7 @@ import { InboxRuleHandler } from "../../mail/model/InboxRuleHandler.js"
 import { Router, ScopedRouter, ThrottledRouter } from "../../gui/ScopedRouter.js"
 import { ShareableGroupType } from "../../sharing/GroupUtils.js"
 import { KdfPicker } from "../../misc/KdfPicker.js"
+import { DomainConfigProvider } from "../common/DomainConfigProvider.js"
 
 assertMainOrNode()
