+++ src/misc/2fa/SecondFactorUtils.ts
@@ -1,4 +1,4 @@
-import { DomainConfigProvider } from "../../api/common/Env.js"
+import { DomainConfigProvider } from "../../api/common/DomainConfigProvider.js"
 
 export function appIdToLoginDomain(appId: string, domainConfigProvider: DomainConfigProvider): string {
 	// If it's legacy U2F key, get domain from before the path part. Otherwise it's just a domain.
