+++ src/types.d.ts
@@ -31,18 +31,38 @@ declare type Base32 = string
 declare type EnvMode = "Browser" | "App" | "Test" | "Playground" | "Desktop" | "Admin"
 declare type PlatformId = "ios" | "android" | "darwin" | "linux" | "win32"
 
+/**
+ * Different parameters based on the domain the app is running on.
+ */
 type DomainConfig = {
+	/** Whether it is a well-known domain provided by us. */
 	firstPartyDomain: boolean
-	/** Important! You probably do not want to use it directly but rather through the accessor function */
+	/**
+	 *  What URL should be used for REST requests.
+	 * Important! You probably do not want to use it directly but rather through the accessor function
+	 */
 	apiUrl: string
+	/**
+	 * Which URL should be opened for Webauthn flow on desktop for keys associated with our current domain (tuta.com).
+	 */
 	webauthnUrl: string
+	/**
+	 * Which URL should b opened for Webauthn flow on desktop for keys associated with our legacy domain (tutanota.com)
+	 */
 	legacyWebauthnUrl: string
+	/** Same as {@link webauthnUrl} but for mobile apps. */
 	webauthnMobileUrl: string
+	/** Same as {@link legacyWebauthnUrl} but for mobile apps. */
 	legacyWebauthnMobileUrl: string
+	/** Which URL should be opened for the credit card payment flow. */
 	paymentUrl: string
+	/** Our current Relying Party ID to register the keys for. Superdomain of our domains. */
 	webauthnRpId: string
+	/** URL for the legacy U2F API. */
 	u2fAppId: string
+	/** Which URL to use to build the gift card sharing URL. */
 	giftCardBaseUrl: string
+	/** Which URL to use to build the referral URL. */
 	referralBaseUrl: string
 }
